self-review/testing cleanup
