+++ src/components/icons/google.tsx
@@ -16,7 +16,7 @@ export function Google({ className }: GoogleProps) {
       <g
         id="Icons"
         stroke="none"
-        stroke-width="1"
+        strokeWidth="1"
         fill="none"
         fill-rule="evenodd"
       >
